rebase and de-c#ipify
