+++ include/prefix_data.hpp
@@ -5,9 +5,6 @@
 #include "nlohmann/json.hpp"
 
 #include "util.hpp"
-#include "path.hpp"
-
-namespace fs = std::filesystem;
 
 namespace mamba
 {
@@ -15,17 +12,7 @@ namespace mamba
     {
     public:
 
-        PackageRecord(const nlohmann::json&& j)
-            : json(std::move(j))
-        {
-            name = json["name"];
-            version = json["version"];
-            build = json["build"];
-            build_number = json["build_number"];
-            channel = json["channel"];
-            subdir = json["subdir"];
-            fn = json["fn"];
-        }
+        PackageRecord(const nlohmann::json&& j);
 
         std::string name;
         std::string version;
@@ -42,46 +29,13 @@ namespace mamba
     class PrefixData
     {
     public:
-        // PrefixData(const fs::path& prefix_path)
-        //     : m_prefix_path(prefix_path)
-        // {
-        // }
+        using package_map = std::unordered_map<std::string, PackageRecord>;
 
-        // constructor for Python
-        PrefixData(const std::string& prefix_path)
-            : m_prefix_path(fs::path(prefix_path))
-        {
-        }
+        PrefixData(const std::string& prefix_path);
 
-        void load()
-        {
-            auto conda_meta_dir = m_prefix_path / "conda-meta";
-            if (path::lexists(conda_meta_dir))
-            {
-                for(auto& p: fs::directory_iterator(conda_meta_dir))
-                {
-                    if (ends_with(p.path().c_str(), ".json"))
-                    {
-                        load_single_record(p.path());
-                    }
-                }
-            }
-        }
-
-        const std::unordered_map<std::string, PackageRecord>& records() const
-        {
-            return m_package_records;
-        }
-
-        void load_single_record(const fs::path& path)
-        {
-            LOG(INFO) << "Loading single package record: " << path;
-            std::ifstream infile(path);
-            nlohmann::json j;
-            infile >> j;
-            auto prec = PackageRecord(std::move(j));
-            m_package_records.insert({prec.name, std::move(prec)});
-        }
+        void load();
+        const package_map& records() const;
+        void load_single_record(const fs::path& path);
 
         std::unordered_map<std::string, PackageRecord> m_package_records;
         fs::path m_prefix_path;
